moving code around to make root.go more spacious
Also worked on having go.mod reflect true repository name.
