+++ main.go
@@ -1,10 +1,9 @@
 /*
 Copyright © 2024 NAME HERE <EMAIL ADDRESS>
-
 */
 package main
 
-import "github.com/kuhlman-labs/gh-commit-remap/cmd"
+import "github.com/mona-actions/gh-commit-remap/cmd"
 
 func main() {
 	cmd.Execute()
